refactor kymaconfig pkg
