+++ internal/cmdcommon/extension_test.go
@@ -19,7 +19,7 @@ import (
 func TestListFromCluster(t *testing.T) {
 	t.Run("list extensions from cluster", func(t *testing.T) {
 		kubeClientConfig := &KubeClientConfig{
-			KubeClient: &fake.KubeClient{
+			kubeClient: &fake.KubeClient{
 				TestKubernetesInterface: k8s_fake.NewSimpleClientset(
 					fixTestExtensionConfigmap("test-1"),
 					fixTestExtensionConfigmap("test-2"),
@@ -46,7 +46,7 @@ func TestListFromCluster(t *testing.T) {
 
 	t.Run("extensions duplications warning", func(t *testing.T) {
 		kubeClientConfig := &KubeClientConfig{
-			KubeClient: &fake.KubeClient{
+			kubeClient: &fake.KubeClient{
 				TestKubernetesInterface: k8s_fake.NewSimpleClientset(
 					fixTestExtensionConfigmapWithCMName("test-1", "test-1"),
 					fixTestExtensionConfigmapWithCMName("test-2", "test-1"),
@@ -76,7 +76,7 @@ func TestListFromCluster(t *testing.T) {
 
 	t.Run("missing rootCommand error", func(t *testing.T) {
 		kubeClientConfig := &KubeClientConfig{
-			KubeClient: &fake.KubeClient{
+			kubeClient: &fake.KubeClient{
 				TestKubernetesInterface: k8s_fake.NewSimpleClientset(
 					&corev1.ConfigMap{
 						ObjectMeta: v1.ObjectMeta{
@@ -106,7 +106,7 @@ func TestListFromCluster(t *testing.T) {
 
 	t.Run("skip optional fields", func(t *testing.T) {
 		kubeClientConfig := &KubeClientConfig{
-			KubeClient: &fake.KubeClient{
+			kubeClient: &fake.KubeClient{
 				TestKubernetesInterface: k8s_fake.NewSimpleClientset(
 					&corev1.ConfigMap{
 						ObjectMeta: v1.ObjectMeta{
