+++ internal/cmdcommon/kubeconfig.go
@@ -11,41 +11,40 @@ import (
 
 // KubeClientConfig allows to setup kubeconfig flag and use it to create kube.Client
 type KubeClientConfig struct {
-	KubeClient    kube.Client
-	KubeClientErr error
+	kubeClient    kube.Client
+	kubeClientErr error
 }
 
-func newKubeClientConfig(cmd *cobra.Command) *KubeClientConfig {
+func newKubeClientConfig() *KubeClientConfig {
 	cfg := &KubeClientConfig{}
-	cfg.addFlag(cmd)
 	cfg.complete()
 
 	return cfg
 }
 
+func AddCmdPersistentKubeconfigFlag(cmd *cobra.Command) {
+	// this flag is not operational. it's only to print help description and help cobra with validation
+	_ = cmd.PersistentFlags().String("kubeconfig", "", "Path to the Kyma kubeconfig file")
+}
+
 func (kcc *KubeClientConfig) GetKubeClient() (kube.Client, error) {
-	return kcc.KubeClient, kcc.KubeClientErr
+	return kcc.kubeClient, kcc.kubeClientErr
 }
 
 func (kcc *KubeClientConfig) GetKubeClientWithClierr() (kube.Client, clierror.Error) {
-	if kcc.KubeClientErr != nil {
-		return nil, clierror.Wrap(kcc.KubeClientErr,
+	if kcc.kubeClientErr != nil {
+		return nil, clierror.Wrap(kcc.kubeClientErr,
 			clierror.New("failed to create cluster connection", "Make sure that kubeconfig is proper."),
 		)
 	}
 
-	return kcc.KubeClient, nil
-}
-
-func (kcc *KubeClientConfig) addFlag(cmd *cobra.Command) {
-	// this flag is not operational. it's only to print help description and help cobra with validation
-	_ = cmd.PersistentFlags().String("kubeconfig", "", "Path to the Kyma kubeconfig file")
+	return kcc.kubeClient, nil
 }
 
 func (kcc *KubeClientConfig) complete() {
 	kubeconfigPath := getKubeconfigPath()
 
-	kcc.KubeClient, kcc.KubeClientErr = kube.NewClient(kubeconfigPath)
+	kcc.kubeClient, kcc.kubeClientErr = kube.NewClient(kubeconfigPath)
 }
 
 // search os.Args manually to find if user pass --kubeconfig path and return its value
