+++ internal/cmdcommon/kymaconfig.go
@@ -2,8 +2,6 @@ package cmdcommon
 
 import (
 	"context"
-
-	"github.com/spf13/cobra"
 )
 
 // KymaConfig contains data common for all subcommands
@@ -14,12 +12,12 @@ type KymaConfig struct {
 	Ctx context.Context
 }
 
-func NewKymaConfig(cmd *cobra.Command) *KymaConfig {
+func NewKymaConfig() *KymaConfig {
 	ctx := context.Background()
 
 	kymaConfig := &KymaConfig{}
 	kymaConfig.Ctx = ctx
-	kymaConfig.KubeClientConfig = newKubeClientConfig(cmd)
+	kymaConfig.KubeClientConfig = newKubeClientConfig()
 	kymaConfig.KymaExtensionsConfig = newExtensionsConfig(kymaConfig)
 
 	return kymaConfig
